XMLHttpRequest/tests/submissions/Opera/getallresponseheaders-status.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: getAllResponseHeaders() excludes status</title>
    <script src="/resources/testharness.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var client = new XMLHttpRequest()
        client.onreadystatechange = function() {
          test.step(function() {
            var headers = client.getAllResponseHeaders().toLowerCase()
            if(client.readyState == 1) {
              assert_equals(headers, "")
            }
            if(client.readyState > 1) {
              assert_false(headers.indexOf("200 ok") != -1)
              assert_false(headers.indexOf("http/1.") != -1)
            }
            if(client.readyState == 4)
              test.done()
          })
        }
        client.open("GET", "resources/headers.php")
        client.send(null)
      })
  </script>
 </body>
</html>
